end detectType and convert origin type

+++ 06/ex00/ScalarConverter.h
@@ -6,7 +6,7 @@
 /*   By: sunko <sunko@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/08 19:30:48 by sunko             #+#    #+#             */
-/*   Updated: 2024/01/09 11:35:21 by sunko            ###   ########.fr       */
+/*   Updated: 2024/01/09 16:09:27 by sunko            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ class ScalarConverter
 		std::string	trimWhiteSpace(const std::string& s);
 		int			detectType(std::string s);
 		bool		tryParseChar(const std::string& s);
-		bool		tryParseInt(const std::string&s, int& result);
+		bool		tryParseInt(const std::string&s, long long& result);
 	private:
 		ScalarConverter();
 		ScalarConverter(const ScalarConverter& rhs);
